refactor(filters): move package `partitions` to `filters/partitioning`
