+++ filters/sobel/concurrent.go
@@ -5,7 +5,7 @@ import (
 	"math"
 	"sync"
 
-	"dougdomingos.com/image-filters/partitions"
+	"dougdomingos.com/image-filters/filters/partitioning"
 	"dougdomingos.com/image-filters/utils"
 )
 
@@ -15,8 +15,8 @@ import (
 func concurrentSobel(img *image.RGBA) {
 	var (
 		bounds      = img.Bounds()
-		numWorkers  = partitions.GetNumberOfWorkers(bounds)
-		imageStrips = partitions.GetVerticalPartitions(bounds, numWorkers)
+		numWorkers  = partitioning.GetNumberOfWorkers(bounds)
+		imageStrips = partitioning.GetVerticalPartitions(bounds, numWorkers)
 		wg          sync.WaitGroup
 	)
